+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/DeployModule.java
@@ -3,46 +3,32 @@ package com.concordium.sdk.transactions;
 import com.concordium.sdk.transactions.smartcontracts.WasmModule;
 import com.concordium.sdk.types.UInt64;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NonNull;
 
 @Getter
+@EqualsAndHashCode(callSuper = true)
 public class DeployModule extends Payload {
     /**
      * A compiled Smart Contract Module in WASM with source and version.
      */
     private final WasmModule module;
 
-    private final UInt64 maxEnergyCost;
 
-    private DeployModule(@NonNull final WasmModule module, @NonNull final UInt64 maxEnergyCost) {
+    private DeployModule(@NonNull final WasmModule module) {
         this.module = module;
-        this.maxEnergyCost = maxEnergyCost;
     }
 
     /**
      * It creates a new instance of the DeployModule class.
      *
      * @param module        The module to be deployed.
-     * @param maxEnergyCost The maximum amount of energy that can be consumed by the contract.
      * @return A new DeployModule object.
      */
     @Builder
-    static DeployModule createNew(final WasmModule module, final UInt64 maxEnergyCost) {
-        return new DeployModule(module, maxEnergyCost);
-    }
-
-    /**
-     * This function returns the type of payload that this class represents.
-     */
-    @Override
-    public PayloadType getType() {
-        return PayloadType.DEPLOY_MODULE;
-    }
-
-    @Override
-    UInt64 getTransactionTypeCost() {
-        return this.maxEnergyCost;
+    static DeployModule createNew(final WasmModule module) {
+        return new DeployModule(module);
     }
 
     @Override
@@ -51,7 +37,7 @@ public class DeployModule extends Payload {
     }
 
     @Override
-    public byte[] getTransactionPayloadBytes() {
+    protected byte[] getRawPayloadBytes() {
         return module.getBytes();
     }
 }