+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/EncryptedTransferWithMemo.java
@@ -18,22 +18,12 @@ public final class EncryptedTransferWithMemo extends Payload {
     }
 
     @Override
-    public PayloadType getType() {
-        return PayloadType.ENCRYPTED_TRANSFER_WITH_MEMO;
-    }
-
-    @Override
-    UInt64 getTransactionTypeCost() {
-        return TransactionTypeCost.ENCRYPTED_TRANSFER_WITH_MEMO.getValue();
-    }
-
-    @Override
     public TransactionType getTransactionType() {
         return TransactionType.ENCRYPTED_TRANSFER_WITH_MEMO;
     }
 
     @Override
-    public byte[] getTransactionPayloadBytes() {
+    protected byte[] getRawPayloadBytes() {
         return payload.getBytes();
     }