+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/ReceiveName.java
@@ -2,6 +2,7 @@ package com.concordium.sdk.transactions;
 
 import com.concordium.sdk.types.UInt16;
 import com.fasterxml.jackson.annotation.JsonCreator;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 import lombok.val;
@@ -16,6 +17,7 @@ import java.nio.ByteBuffer;
 
 @Getter
 @ToString
+@EqualsAndHashCode
 public final class ReceiveName {
     /**
      * The name of the contract.