+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransferScheduleTransaction.java
@@ -2,6 +2,8 @@ package com.concordium.sdk.transactions;
 
 import com.concordium.sdk.exceptions.TransactionCreationException;
 import com.concordium.sdk.types.AccountAddress;
+import com.concordium.sdk.types.UInt16;
+import com.concordium.sdk.types.UInt64;
 import lombok.*;
 
 
@@ -20,7 +22,13 @@ public class TransferScheduleTransaction extends AccountTransaction {
             @NonNull final AccountNonce nonce,
             @NonNull final Expiry expiry,
             @NonNull final TransactionSigner signer) {
-        super(sender, nonce, expiry, signer, TransferSchedule.createNew(to, schedule));
+        super(sender, nonce, expiry, signer, TransferSchedule.createNew(to, schedule),getCost(schedule));
+    }
+
+    private static UInt64 getCost(@NonNull Schedule[] schedule) {
+        UInt16 scheduleLen = UInt16.from(schedule.length);
+        val maxEnergyCost = UInt64.from(scheduleLen.getValue()).getValue() * (300 + 64);
+        return UInt64.from(maxEnergyCost);
     }
 
     /**