+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransferToPublic.java
@@ -37,27 +37,13 @@ public final class TransferToPublic extends Payload {
      */
     private final SecToPubAmountTransferProof proof;
 
-
-    /**
-     * This function returns the type of the payload.
-     */
-    @Override
-    public PayloadType getType() {
-        return PayloadType.TRANSFER_TO_PUBLIC;
-    }
-
-    @Override
-    UInt64 getTransactionTypeCost() {
-        return TransactionTypeCost.TRANSFER_TO_PUBLIC.getValue();
-    }
-
     @Override
     public TransactionType getTransactionType() {
         return TransactionType.TRANSFER_TO_PUBLIC;
     }
 
     @Override
-    public byte[] getTransactionPayloadBytes() {
+    protected byte[] getRawPayloadBytes() {
         val proofBytes = this.proof.getBytes();
         val remainingAmountBytes = this.remainingAmount.getBytes();
         val buffer = ByteBuffer.allocate(remainingAmountBytes.length