+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/UpdateContractTransaction.java
@@ -14,40 +14,30 @@ import lombok.*;
 @ToString(callSuper = true)
 public class UpdateContractTransaction extends AccountTransaction {
     private UpdateContractTransaction(
-            @NonNull final UpdateContractPayload payload,
+            @NonNull final UpdateContract payload,
             @NonNull final AccountAddress sender,
             @NonNull final AccountNonce nonce,
             @NonNull final Expiry expiry,
             @NonNull final TransactionSigner signer,
             @NonNull final UInt64 maxEnergyCost) {
-        super(sender, nonce, expiry, signer, UpdateContract.createNew(payload, maxEnergyCost));
-    }
-
-    private UpdateContractTransaction(
-            final @NonNull TransactionHeader header,
-            final @NonNull TransactionSignature signature,
-            final @NonNull UpdateContractPayload payload) {
-        super(header,
-                signature,
-                TransactionType.DEPLOY_MODULE,
-                payload.getBytes());
+        super(sender, nonce, expiry, signer, payload, maxEnergyCost);
     }
 
     /**
      * Creates a new instance of {@link UpdateContractTransaction}.
      *
-     * @param payload       {@link UpdateContractPayload} of the transaction.
+     * @param payload       {@link UpdateContract} of the transaction.
      * @param sender        Sender ({@link AccountAddress}) of this Transaction.
      * @param nonce         Account {@link com.concordium.sdk.types.Nonce} Of the Sender Account.
      * @param expiry        {@link Expiry} of this transaction.
      * @param signer        {@link Signer} of this transaction.
-     * @param maxEnergyCost Allowed energy of the transaction.
+     * @param maxEnergyCost the maximum energy allowed to spend on this transaction.
      * @throws TransactionCreationException On failure to create the Transaction from input params.
      *                                      Ex when any of the input param is NULL.
      */
     @Builder(builderClassName = "UpdateContractTransactionBuilder")
     public static UpdateContractTransaction from(
-            final UpdateContractPayload payload,
+            final UpdateContract payload,
             final AccountAddress sender,
             final AccountNonce nonce,
             final Expiry expiry,
@@ -60,25 +50,4 @@ public class UpdateContractTransaction extends AccountTransaction {
         }
     }
 
-    /**
-     * Creates a new instance of {@link UpdateContractTransaction}.
-     * Using {@link TransactionHeader}, {@link TransactionSignature} and Payload {@link UpdateContractPayload}.
-     *
-     * @param header    {@link TransactionHeader}.
-     * @param signature {@link TransactionSignature}.
-     * @param payload   {@link UpdateContractPayload} Payload for this transaction.
-     * @throws TransactionCreationException On failure to create the Transaction from input params.
-     *                                      Ex when any of the input param is NULL.
-     */
-    @Builder(builderMethodName = "builderBlockItem", builderClassName = "UpdateContractBlockItemBuilder")
-    public static UpdateContractTransaction from(
-            final TransactionHeader header,
-            final TransactionSignature signature,
-            final UpdateContractPayload payload) {
-        try {
-            return new UpdateContractTransaction(header, signature, payload);
-        } catch (NullPointerException nullPointerException) {
-            throw TransactionCreationException.from(nullPointerException);
-        }
-    }
 }