+++ concordium-sdk/src/test/java/com/concordium/sdk/transactions/UpdateContractTest.java
@@ -21,28 +21,27 @@ public class UpdateContractTest {
     @Test
     public void updateContractTest() {
         byte[] emptyArray = new byte[0];
-        val transfer = UpdateContract.createNew(
-                        UpdateContractPayload.from(0, ContractAddress.from(81, 0), "CIS2-NFT", "mint", emptyArray), UInt64.from(3000))
-                .withHeader(TransactionHeader
-                        .builder()
-                        .sender(AccountAddress.from("3JwD2Wm3nMbsowCwb1iGEpnt47UQgdrtnq2qT6opJc3z2AgCrc"))
-                        .accountNonce(Nonce.from(78910))
-                        .expiry(UInt64.from(123456))
-                        .build())
-                .signWith(
-                        TransactionSigner.from(
-                                SignerEntry.from(Index.from(0), Index.from(0),
-                                        ED25519SecretKey.from("7100071c835a0a35e86dccba7ee9d10b89e36d1e596771cdc8ee36a17f7abbf2")),
-                                SignerEntry.from(Index.from(0), Index.from(1),
-                                        ED25519SecretKey.from("cd20ea0127cddf77cf2c20a18ec4516a99528a72e642ac7deb92131a9d108ae9"))
-                        )
-                );
-        val transferBytesLength = transfer.getBytes().length;
-        assertEquals(42, transferBytesLength);
+        UpdateContract updateContractPayload = UpdateContract.from(0, ContractAddress.from(81, 0), "CIS2-NFT", "mint", emptyArray);
+        UpdateContractTransaction tx = TransactionFactory.newUpdateContract()
+                .payload(updateContractPayload)
+                .sender(AccountAddress.from("3JwD2Wm3nMbsowCwb1iGEpnt47UQgdrtnq2qT6opJc3z2AgCrc"))
+                .nonce(AccountNonce.from(78910))
+                .expiry(Expiry.from(123456))
+                .maxEnergyCost(UInt64.from(3000))
+                .signer(TransactionSigner.from(
+                        SignerEntry.from(Index.from(0), Index.from(0),
+                                ED25519SecretKey.from("7100071c835a0a35e86dccba7ee9d10b89e36d1e596771cdc8ee36a17f7abbf2")),
+                        SignerEntry.from(Index.from(0), Index.from(1),
+                                ED25519SecretKey.from("cd20ea0127cddf77cf2c20a18ec4516a99528a72e642ac7deb92131a9d108ae9"))
+                ))
+                .build();
+        val payload = tx.getPayload();
+        val txLength = payload.getBytes().length;
+        assertEquals(42, txLength);
 
-        val transferDataToSign = transfer.getDataToSign();
+        val transferDataToSign = payload.getDataToSign();
         assertEquals("02b15fc092aa1e63035a18cb1b4f62a9d2a6186bdb6147f85525bcb8771ab5b7", Hex.encodeHexString(transferDataToSign));
-        val blockItem = transfer.toAccountTransaction();
+        val blockItem = payload.toAccountTransaction();
 
         val blockItemBytes = blockItem.getBytes();
         assertArrayEquals(EXPECTED_BLOCK_ITEM_UPDATE_CONTRACT_DATA_BYTES, TestUtils.signedByteArrayToUnsigned(blockItemBytes));